// src/app/interfaces/asset-interface.ts
export interface Asset {
  ['created-at-round']: number;
  deleted: boolean;
  index: number;
  params: {
    clawback: string;
    creator: string;
    decimals: number;
    ['default-frozen']: boolean;
    freeze: string;
    manager: string;
    ['metadata-hash']: string;
    name: string;
    ['name-b64']: string;
    reserve: string;
    total: number;
    ['unit-name']: string;
    ['unit-name-b64']: string;
    url: string;
    ['url-b64']: string;
  };
}